refactor[#17] : create...supabase shared/model 슬라이스로 이동
