+++ blog/src/shared/model/index.ts
@@ -1,0 +1,2 @@
+export * from "./createServerSupabase";
+export * from "./createBrowserSupabase";
