+++ blog/src/shared/utils/index.ts
@@ -1,6 +1,4 @@
-export * from "./createServerSupabase";
 export * from "./supabaseMiddleware";
-export * from "./createBrowserSupabase";
 export * from "./camelToSnake";
 export * from "./snakeToCamel";
 export * from "./createPostgressErrorResponse";
